refactor: move dedicated components to `features/`
