+++ src/features/contact_me/MessageMe.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import MessageMeForm from "@/components/contact_me/MessageMeForm";
 import { Button } from "@/components/shadcn_ui/button";
 import {
 	Dialog,
@@ -18,6 +17,7 @@ import {
 	DrawerTitle,
 	DrawerTrigger,
 } from "@/components/shadcn_ui/drawer";
+import MessageMeForm from "@/features/contact_me/MessageMeForm";
 import { Breakpoints, useMediaQuery } from "@/hooks/use-media-query";
 import { cn } from "@/lib/utils";
 import React, { useEffect, useState } from "react";
